requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.0
google-generativeai>=0.3.0
